# Reduce calls to ToArray in LightweightObservableBase PublishNext