## What does the pull request do?
<!--- Give a bit of background on the PR here, together with links to with related issues etc. -->
LightweightObservableBase PublishNext currently calls ToArray to make a protected set of observers but in the vast majority of cases there is only one observer.

## What is the current behavior?
<!--- If the PR is a fix, describe the current incorrect behavior, otherwise delete this section. -->
An array is allocated on every call

## What is the updated/expected behavior with this PR?
<!--- Describe how to test the PR. -->
No allocation on the common case of just one observer

## How was the solution implemented (if it's not obvious)?
<!--- Include any information that might be of use to a reviewer here. -->


## Checklist

- [ ] Added unit tests (if possible)?
- [ ] Added XML documentation to any related classes?
- [ ] Consider submitting a PR to https://github.com/AvaloniaUI/Avaloniaui.net with user documentation

## Breaking changes
<!--- List any breaking changes here. When the PR is merged please add an entry to https://github.com/AvaloniaUI/Avalonia/wiki/Breaking-Changes -->


## Fixed issues
<!--- If the pull request fixes issue(s) list them like this: 
Fixes #123
Fixes #456
-->