@MarchingCube 2019-11-20
Could you write a benchmark test for this?