@ahopper 2019-11-20
I'll try and write a benchmark.  In looking deeper into this I'm intrigued by all the threading stuff in this class, I'm yet to convince myself of the need for it.  Longer term a better solution would avoid creating the _observers collection unless needed but I need to understand the need for the threading stuff first.