@MarchingCube 2019-11-20
I think this is inspired by `System.Reactive` classes which use similar constructs a lot, mostly because this object can be accessed from any thread. Not sure if this is the case in many instances for us, since all properties check if we are on UI thread anyway. Also I am no so sure if deallocating whole storage when removing last subscriber is actually beneficial. Since we end up constantly reallocating when somebody constantly subscribes and unsubscribes.