@ahopper 2019-11-20
I added to the binding benchmarks as that is what I was investigating when I spotted it.
The second benchmark creates a binding then updates it 60 times.
```
before
|                          Method |      Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------------- |----------:|----------:|----------:|-------:|------:|------:|----------:|
|       TwoWayBinding_Via_Binding |  11.90 us | 0.1823 us | 0.1616 us | 0.9766 |     - |     - |   3.03 KB |
| UpdateTwoWayBinding_Via_Binding | 245.60 us | 4.7302 us | 5.0613 us | 9.2773 |     - |     - |  28.94 KB |

after

|                          Method |      Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------------- |----------:|----------:|----------:|-------:|------:|------:|----------:|
|       TwoWayBinding_Via_Binding |  11.68 us | 0.2262 us | 0.2861 us | 0.9613 |     - |     - |   2.97 KB |
| UpdateTwoWayBinding_Via_Binding | 226.35 us | 4.4368 us | 4.1502 us | 7.5684 |     - |     - |  23.34 KB |
```